requirements.txt
flask==3.0.2
werkzeug==3.0.1
openai-whisper==20231117
google-generativeai==0.3.2
moviepy==1.0.3
numpy==1.26.4
gTTS==2.5.1
ratelimit==2.2.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23 